# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ProyectoAtlas.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/projects/{projectSlug}/audit-events": {
      "get": {
        "tags": [
          "ProjectAuditEvents"
        ],
        "parameters": [
          {
            "name": "projectSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/AuditEntityType"
            }
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/AuditAction"
            }
          },
          {
            "name": "occurredFromUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "occurredToUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProjectAuditEventsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{projectSlug}/release-notes": {
      "post": {
        "tags": [
          "ProjectReleaseNotes"
        ],
        "parameters": [
          {
            "name": "projectSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectReleaseNotesCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectReleaseNotesCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectReleaseNotesCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Documentation"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "ProjectReleaseNotes"
        ],
        "parameters": [
          {
            "name": "projectSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DocumentationStatus"
            }
          },
          {
            "name": "area",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DocumentationArea"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProjectDocumentationsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{projectSlug}/release-notes/{slug}": {
      "get": {
        "tags": [
          "ProjectReleaseNotes"
        ],
        "parameters": [
          {
            "name": "projectSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Documentation"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "ProjectReleaseNotes"
        ],
        "parameters": [
          {
            "name": "projectSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectReleaseNotesCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectReleaseNotesCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectReleaseNotesCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Documentation"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ProjectReleaseNotes"
        ],
        "parameters": [
          {
            "name": "projectSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": { }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{slug}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": { }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "required": [
          "statusCode",
          "message",
          "code"
        ],
        "type": "object",
        "properties": {
          "statusCode": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "HTTP status code returned by the API.",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "description": "Developer-oriented message that explains the error."
          },
          "code": {
            "type": "string",
            "description": "Stable application error code intended for frontend flows."
          }
        },
        "example": {
          "statusCode": 409,
          "message": "Project slug 'proyecto-atlas' already exists.",
          "code": "PROJECT_SLUG_CONFLICT"
        }
      },
      "AuditAction": {
        "enum": [
          "Created",
          "Updated",
          "Deleted"
        ],
        "type": "string"
      },
      "AuditEntityType": {
        "enum": [
          "Project",
          "Documentation"
        ],
        "type": "string"
      },
...
      "CreateProjectDocumentationTag": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "CreateProjectReleaseNotesCommand": {
        "required": [
          "title",
          "contentMarkdown",
          "sortOrder",
          "status",
          "area"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "contentMarkdown": {
            "type": "string"
          },
          "sortOrder": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentationStatus"
          },
          "area": {
            "$ref": "#/components/schemas/DocumentationArea"
          },
          "tags": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/CreateProjectDocumentationTag"
            }
          }
        }
      },
      "Documentation": {
        "required": [
          "projectId",
          "title",
          "contentMarkdown",
          "sortOrder",
          "kind",
          "status",
          "area"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "contentMarkdown": {
            "type": "string"
          },
          "sortOrder": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "kind": {
            "$ref": "#/components/schemas/DocumentationKind"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentationStatus"
          },
          "area": {
            "$ref": "#/components/schemas/DocumentationArea"
          },
          "tags": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/DocumentationTag"
            }
          },
          "faqItems": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/DocumentationFaqItem"
            }
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "id": "d7f2cb87-86bc-4c25-bb77-27e4f7c38b67",
          "projectId": "8b658c72-8f6f-4fef-9d65-f2fa6eb60bd7",
          "title": "Getting Started",
          "slug": "getting-started",
          "contentMarkdown": "# Proyecto Atlas\n\nGuia inicial para levantar el backend localmente.",
          "sortOrder": 1,
          "kind": "Page",
          "status": "Draft",
          "createdAtUtc": "2026-03-24T18:35:00Z",
          "updatedAtUtc": "2026-03-24T18:35:00Z"
        }
      },
...
      "DocumentationArea": {
        "enum": [
          "Product",
          "Architecture",
          "Backend",
          "Frontend",
          "Operations",
          "DevOps",
          "Security",
          "UX"
        ],
        "type": "string"
      },
...
      "DocumentationKind": {
        "enum": [
          "Page",
          "Decision",
          "Note",
          "FAQ",
          "ReleaseNotes"
        ],
        "type": "string"
      },
...
      "DocumentationStatus": {
        "enum": [
          "Draft",
          "Published",
          "Archived"
        ],
        "type": "string"
      },
...
      "ListProjectAuditEventsResponse": {
        "required": [
          "items"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditEvent"
            }
          }
        },
        "example": {
          "items": [
            {
              "id": "5f57b8bb-a4f7-4b6a-aa5f-f5624d2af6f7",
              "projectId": "8b658c72-8f6f-4fef-9d65-f2fa6eb60bd7",
              "documentationId": null,
              "entityType": "Project",
              "entityId": "8b658c72-8f6f-4fef-9d65-f2fa6eb60bd7",
              "entitySlug": "proyecto-atlas",
              "entityTitle": "Proyecto Atlas",
              "action": "Updated",
              "occurredAtUtc": "2026-04-16T12:00:00Z"
            }
          ]
        }
      },
      "ListProjectDocumentationsResponse": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalPages",
          "totalItems"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Documentation"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalItems": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        },
        "example": {
          "items": [
            {
              "id": "d7f2cb87-86bc-4c25-bb77-27e4f7c38b67",
              "projectId": "8b658c72-8f6f-4fef-9d65-f2fa6eb60bd7",
              "title": "Getting Started",
              "slug": "getting-started",
              "contentMarkdown": "# Proyecto Atlas\n\nGuia inicial para levantar el backend localmente.",
              "sortOrder": 1,
              "kind": "Page",
              "status": "Draft",
              "createdAtUtc": "2026-03-24T18:35:00Z",
              "updatedAtUtc": "2026-03-24T18:35:00Z"
            }
          ],
          "page": 1,
          "pageSize": 10,
          "totalPages": 1,
          "totalItems": 1
        }
      },
...
      "Project": {
        "required": [
          "title",
          "description",
          "repositoryUrl",
          "color"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "repositoryUrl": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "links": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ProjectLink"
            }
          }
        },
        "example": {
          "id": "8b658c72-8f6f-4fef-9d65-f2fa6eb60bd7",
          "title": "Proyecto Atlas",
          "description": "Backend for project documentation based on markdown",
          "repositoryUrl": "https://github.com/matigaleanodev/proyecto-atlas-api",
          "color": "#1E293B",
          "slug": "proyecto-atlas",
          "links": [
            {
              "id": "4d987fe2-3489-46db-9b83-e493750a8d8e",
              "projectId": "8b658c72-8f6f-4fef-9d65-f2fa6eb60bd7",
              "title": "Repository",
              "url": "https://github.com/matigaleanodev/proyecto-atlas-api",
              "description": "Main source code",
              "sortOrder": 1,
              "kind": "Repository"
            }
          ],
          "createdAtUtc": "2026-03-24T18:30:00Z",
          "updatedAtUtc": "2026-03-24T18:30:00Z"
        }
      },
...
      "UpdateProjectCommand": {
        "type": "object",
        "properties": {
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "repositoryUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "color": {
            "type": [
              "null",
              "string"
            ]
          },
          "links": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/UpdateProjectLink"
            }
          }
        },
        "example": {
          "title": "Atlas Platform",
          "description": "Updated backend for project documentation",
          "repositoryUrl": "https://github.com/matigaleanodev/atlas-platform",
          "color": "#0F172A",
          "links": [
            {
              "title": "Monitoring",
              "url": "https://grafana.example.com/atlas-platform",
              "description": "Operational dashboards",
              "sortOrder": 1,
              "kind": "Monitoring"
            }
          ]
        }
      },
...
      "UpdateProjectReleaseNotesCommand": {
        "required": [
          "title",
          "contentMarkdown",
          "sortOrder",
          "status"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "contentMarkdown": {
            "type": [
              "null",
              "string"
            ]
          },
          "sortOrder": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DocumentationStatus"
              }
            ]
          },
          "tags": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/UpdateProjectDocumentationTag"
            }
          }
        }
      }
    }
  },
...
}